+++ spec/integration_spec.rb
@@ -17,7 +17,7 @@ describe 'Integration Tests' do
                # no rules
             end
 
-            Sinatra::Bouncer::BasicBouncer::HTTP_METHOD_SYMBOLS.each do |http_method|
+            Sinatra::Bouncer::Rule::HTTP_METHOD_SYMBOLS.each do |http_method|
                send(http_method, route_path) do
                   # whatever
                end
@@ -28,7 +28,7 @@ describe 'Integration Tests' do
       let(:browser) { Rack::Test::Session.new(server_klass) }
 
       it 'should auto-protect all HTTP methods' do
-         Sinatra::Bouncer::BasicBouncer::HTTP_METHOD_SYMBOLS.each do |http_method|
+         Sinatra::Bouncer::Rule::HTTP_METHOD_SYMBOLS.each do |http_method|
             response = browser.send http_method, path
 
             expect(response).to be_forbidden
@@ -39,13 +39,13 @@ describe 'Integration Tests' do
    describe 'rules' do
       it 'should allow access to matching route string' do
          path      = '/admin/dashboard'
-         test_body = 'Test content'
+         test_body = 'Boring is always best.'
 
          server_klass = Class.new Sinatra::Base do
             register Sinatra::Bouncer
 
             rules do
-               can get: path
+               anyone.can get: path
             end
 
             get path do
@@ -63,13 +63,13 @@ describe 'Integration Tests' do
 
       it 'should allow access to matching route string array' do
          paths     = %w[/admin/dashboard /admin/settings]
-         test_body = 'Test content'
+         test_body = 'Boring is always best.'
 
          server_klass = Class.new Sinatra::Base do
             register Sinatra::Bouncer
 
             rules do
-               can get: paths
+               anyone.can get: paths
             end
 
             paths.each do |path|
@@ -91,7 +91,7 @@ describe 'Integration Tests' do
 
       it 'should allow access to matching hash of methods to routes' do
          path      = '/admin/dashboard'
-         test_body = 'Test content'
+         test_body = 'Boring is always best.'
 
          server_klass = Class.new Sinatra::Base do
             register Sinatra::Bouncer
@@ -105,8 +105,8 @@ describe 'Integration Tests' do
             end
 
             rules do
-               can get:  path,
-                   post: path
+               anyone.can get:  path,
+                          post: path
             end
          end
 
@@ -123,7 +123,7 @@ describe 'Integration Tests' do
          expect(response.body).to eq test_body
       end
 
-      it 'should evaludate the rules block once at startup' do
+      it 'should evaluate the rules block once at startup' do
          path = '/admin/dashboard'
 
          runs = 0
@@ -132,13 +132,13 @@ describe 'Integration Tests' do
             register Sinatra::Bouncer
 
             get path do
-               'Test content'
+               'Boring is always best.'
             end
 
             rules do
                runs += 1
 
-               can get: path
+               anyone.can get: path
             end
          end
 
@@ -153,6 +153,106 @@ describe 'Integration Tests' do
          expect(runs).to eq 1
       end
 
+      it 'should accept defining roles' do
+         path = '/admin/dashboard'
+
+         server_klass = Class.new Sinatra::Base do
+            register Sinatra::Bouncer
+
+            get path do
+               'Boring is always best.'
+            end
+
+            role :user do
+               # dummy user session handling for testing - do not do it this way in real life
+               current_user = JSON.parse(request.get_header('X-FAKE-SESSION'))
+
+               !current_user.nil?
+            end
+
+            rules do
+               user.can get: path
+            end
+         end
+
+         browser = Rack::Test::Session.new(server_klass)
+
+         rack_env = {'X-FAKE-SESSION' => {name: 'Michael Bryce'}.to_json}
+         response = browser.get(path, {}, rack_env)
+         expect(response).to be_ok
+      end
+
+      it 'should evaluate both role and rule conditions' do
+         path = '/admin/dashboard'
+
+         server_klass = Class.new Sinatra::Base do
+            register Sinatra::Bouncer
+
+            get path do
+               # whatever
+            end
+
+            role :user do
+               request.get_header('X-CUSTOM-USER') == 'Bryce'
+            end
+
+            rules do
+               user.can_sometimes get: path do
+                  request.get_header('X-CUSTOM-COND') == 'special'
+               end
+            end
+         end
+
+         browser = Rack::Test::Session.new(server_klass)
+
+         rack_env = {'X-CUSTOM-COND' => 'special'}
+         response = browser.get(path, {}, rack_env)
+         expect(response).to be_forbidden
+
+         rack_env = {'X-CUSTOM-USER' => 'Bryce'}
+         response = browser.get(path, {}, rack_env)
+         expect(response).to be_forbidden
+      end
+
+      it 'should evaluate role definitions in the context of the request' do
+         path = '/admin/dashboard'
+
+         server_klass = Class.new Sinatra::Base do
+            register Sinatra::Bouncer
+
+            get path do
+               'Boring is always best.'
+            end
+
+            role :user do
+               !current_user.nil?
+            end
+
+            rules do
+               user.can get: path
+            end
+
+            helpers do
+               # dummy user session handling for testing - do not do it this way in real life
+               def current_user
+                  user = request.get_header('X-FAKE-SESSION')
+
+                  user ? JSON.parse(user) : nil
+               end
+            end
+         end
+
+         browser = Rack::Test::Session.new(server_klass)
+
+         rack_env = {}
+         response = browser.get(path, {}, rack_env)
+         expect(response).to be_forbidden
+
+         rack_env = {'X-FAKE-SESSION' => {name: 'Michael Bryce'}.to_json}
+         response = browser.get(path, {}, rack_env)
+         expect(response).to be_ok
+      end
+
       it 'should evaluate rule conditions in the context of the request' do
          path = '/admin/dashboard'
 
@@ -160,11 +260,11 @@ describe 'Integration Tests' do
             register Sinatra::Bouncer
 
             get path do
-               'Test content'
+               'Boring is always best.'
             end
 
             rules do
-               can_sometimes get: path do
+               anyone.can_sometimes get: path do
                   request.get_header('X-CUSTOM') == 'special'
                end
             end
@@ -172,7 +272,6 @@ describe 'Integration Tests' do
 
          browser = Rack::Test::Session.new(server_klass)
 
-         # first attempt should succeed and flip the has_run variable
          rack_env = {'X-CUSTOM' => 'special'}
          response = browser.get(path, {}, rack_env)
          expect(response).to be_ok
@@ -192,7 +291,7 @@ describe 'Integration Tests' do
             end
 
             rules do
-               can_sometimes get: path do
+               anyone.can_sometimes get: path do
                   !has_run
                end
             end
@@ -216,11 +315,11 @@ describe 'Integration Tests' do
             register Sinatra::Bouncer
 
             get path do
-               'Test content'
+               'Boring is always best.'
             end
 
             rules do
-               can get: path
+               anyone.can get: path
             end
          end
 
@@ -242,7 +341,7 @@ describe 'Integration Tests' do
             end
 
             rules do
-               can head: path
+               anyone.can head: path
             end
          end
 
@@ -264,7 +363,7 @@ describe 'Integration Tests' do
             end
 
             rules do
-               can_sometimes get: path do
+               anyone.can_sometimes get: path do
                   5
                end
             end
@@ -278,7 +377,7 @@ describe 'Integration Tests' do
       end
 
       it 'should NOT complain if a rule returns a falsey non-false' do
-         path = '/admin/dashboard'
+         path = '/blogs/hello-world'
 
          server_klass = Class.new Sinatra::Base do
             register Sinatra::Bouncer
@@ -288,7 +387,7 @@ describe 'Integration Tests' do
             end
 
             rules do
-               can_sometimes get: path do
+               anyone.can_sometimes get: path do
                   nil
                end
             end
@@ -303,46 +402,46 @@ describe 'Integration Tests' do
 
       describe 'wildcards' do
          it 'should match wildcards at the end' do
-            test_body = 'Test content'
+            test_body = 'Boring is always best.'
 
             server_klass = Class.new Sinatra::Base do
                register Sinatra::Bouncer
 
-               get '/admin/dashboard' do
+               get '/blogs/hello-world' do
                   test_body
                end
 
                rules do
-                  can get: '/admin/*'
+                  anyone.can get: '/blogs/*'
                end
             end
 
             browser = Rack::Test::Session.new(server_klass)
 
-            response = browser.get '/admin/dashboard'
+            response = browser.get '/blogs/hello-world'
 
             expect(response).to be_ok
             expect(response.body).to eq test_body
          end
 
          it 'should match wildcards in the middle' do
-            test_body = 'Test content'
+            test_body = 'Boring is always best.'
 
             server_klass = Class.new Sinatra::Base do
                register Sinatra::Bouncer
 
-               get '/admin/dashboard' do
+               get '/blogs/hello-world' do
                   test_body
                end
 
                rules do
-                  can get: '/*/dashboard'
+                  anyone.can get: '/*/hello-world'
                end
             end
 
             browser = Rack::Test::Session.new(server_klass)
 
-            response = browser.get '/admin/dashboard'
+            response = browser.get '/blogs/hello-world'
 
             expect(response).to be_ok
             expect(response.body).to eq test_body